+++ src/main/kotlin/dev/dediamondpro/resourcify/services/modrinth/PartialModrinthProject.kt
@@ -18,10 +18,7 @@
 package dev.dediamondpro.resourcify.services.modrinth
 
 import com.google.gson.annotations.SerializedName
-import dev.dediamondpro.resourcify.services.IGalleryImage
-import dev.dediamondpro.resourcify.services.IProject
-import dev.dediamondpro.resourcify.services.IMember
-import dev.dediamondpro.resourcify.services.IVersion
+import dev.dediamondpro.resourcify.services.*
 import dev.dediamondpro.resourcify.util.*
 import java.awt.Color
 import java.net.URI
@@ -50,6 +47,16 @@ data class PartialModrinthProject(
 
     override fun getName(): String = title
     override fun getId(): String = id
+
+    override fun getType(): ProjectType {
+        return when (projectType) {
+            "resourcepack" -> ProjectType.RESOURCE_PACK
+            "datapack" -> ProjectType.DATA_PACK
+            "shader" -> ProjectType.IRIS_SHADER // Whether it's OptiFine or Iris doesn't matter
+            else -> ProjectType.UNKNOWN
+        }
+    }
+
     override fun getSummary(): String = summary
     override fun getAuthor(): String = author
     override fun getIconUrl(): URI? = iconUrl?.toURIOrNull()